\name{MMBsplines-package}
\alias{MMBsplines-package}
\alias{MMBsplines}
\docType{package}
\title{
Mixed Model B-splines algorithm.
}
\description{
A fast Mixed Model B-splines algorithm in mixed models is implementend. B-splines have local support 
and are computational attractive, because the corresponding matrices are sparse. 
A key element of the algorithm is that the local character of B-splines is preserved. 
The computation time for the algorithm is linear in the number of B-splines. 
}
\details{
\tabular{ll}{
Package: \tab MMBsplines\cr
Type: \tab Package\cr
Version: \tab 1.09\cr
Date: \tab 2015-12-29\cr
License: \tab GPL\cr
}
}
\author{Martin Boer

Maintainer: Martin Boer <martin.boer@wur.nl>
}
\seealso{
 \code{\link{MMBsplines}
}
}

